# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Danshin Genealogy",
        "description": "Доступ ко всем данным только через API токен,\n        в противном случае будет возвращаться ошибка 401.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dates/upcoming": {
            "get": {
                "operationId": "4bbd43443fba95f7662a6111cbb24add",
                "parameters": [
                    {
                        "name": "date",
                        "in": null,
                        "description": "Текущее число в формате гггг-мм-дд",
                        "required": true
                    },
                    {
                        "name": "past_day",
                        "in": null,
                        "description": "Количество предшествующих дней (число от 1 до 30).",
                        "required": true
                    },
                    {
                        "name": "nearest_day",
                        "in": null,
                        "description": "Количество будующих дней (число от 1 до 30).",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список ближайших дат (предыдущие дни, текущий день, будущие дни,\n                    в соответтвии с переданными параметрами)."
                    }
                }
            }
        },
        "/api/dates/notify-all": {
            "get": {
                "operationId": "45b3c2c66eb02d7ca4ea51c0e9b7e61a",
                "parameters": [
                    {
                        "name": "date",
                        "in": null,
                        "description": "Текущее число в формате гггг-мм-дд",
                        "required": true
                    },
                    {
                        "name": "past_day",
                        "in": null,
                        "description": "Количество предшествующих дней (число от 1 до 30).",
                        "required": true
                    },
                    {
                        "name": "nearest_day",
                        "in": null,
                        "description": "Количество будующих дней (число от 1 до 30).",
                        "required": true
                    },
                    {
                        "name": "path_person",
                        "in": null,
                        "description": "Путь к лицу.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Уведомления о ближайших датах направлено."
                    },
                    "500": {
                        "description": "Ошибки при отправлении уведомлений."
                    }
                }
            }
        },
        "/api/dates/subscription/data/{personId}": {
            "get": {
                "operationId": "9d26ed897048996ec9bee9fd69d5e9eb",
                "parameters": [
                    {
                        "name": "personId",
                        "in": null,
                        "description": "Id пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Url телеграм бота и код."
                    },
                    "404": {
                        "description": "Пользователя с переданным id не найден."
                    }
                }
            }
        },
        "/api/dates/subscription/create/{personId}": {
            "get": {
                "operationId": "6ad713b762fe75d26f4366c3346c87e7",
                "responses": {
                    "200": {
                        "description": "Создание подписки."
                    },
                    "404": {
                        "description": "Пользователя с переданным id не найден."
                    }
                }
            }
        },
        "/api/dates/subscription/delete/{personId}": {
            "get": {
                "operationId": "c1c90e721576735eb2b4a9bfc21ef7c9",
                "parameters": [
                    {
                        "name": "personId",
                        "in": null,
                        "description": "Id пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удаление подписки."
                    },
                    "404": {
                        "description": "Пользователя с переданным id не найден."
                    }
                }
            }
        },
        "/api/dates/subscription/exists/{personId}": {
            "get": {
                "operationId": "8673357fc6fe08c765e8fe9b285cc72a",
                "parameters": [
                    {
                        "name": "personId",
                        "in": null,
                        "description": "Id пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Существование подписки."
                    },
                    "404": {
                        "description": "Пользователь с переданным id не найден."
                    }
                }
            }
        },
        "/api/download/people": {
            "get": {
                "operationId": "10b720343e2ee730f9dbc16ac71b7fac",
                "parameters": [
                    {
                        "name": "type",
                        "in": null,
                        "description": "Формат файла. Доступные форматы: pdf.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл, содержащий всех лиц родословной."
                    },
                    "302": {
                        "description": "Не указан формат файла в параметрах запроса."
                    },
                    "404": {
                        "description": "Файл в переданном формате не найден (не поддерживается)."
                    }
                }
            }
        },
        "/api/download/person/{id}": {
            "get": {
                "operationId": "08edc056a1e59b1abb6a825ca6295647",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица.",
                        "required": true
                    },
                    {
                        "name": "type",
                        "in": null,
                        "description": "Формат файла. Доступные форматы: pdf.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл, содержащий лицо с переданным id."
                    },
                    "302": {
                        "description": "Не указан формат файла в параметрах запроса."
                    },
                    "404": {
                        "description": "Лицо с переданным id не найдено или\n                    файл в переданном формате не найден (не поддерживается)."
                    }
                }
            }
        },
        "/api/download/tree/{id}": {
            "get": {
                "operationId": "86f7cfe25414a6c4321f42bca696428b",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица.",
                        "required": true
                    },
                    {
                        "name": "parent_id",
                        "in": null,
                        "description": "Id родителя (запрашиваемого лица) по которому строится древо,\n                    по умолчанию древо строится по отцу. Id родителя должно иметь тип integer."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл в формате SVG, содержащий древо лица с переданным id."
                    },
                    "302": {
                        "description": "Параметры запроса не соответствуют ограничениям."
                    },
                    "404": {
                        "description": "Лицо с переданным id не найдено\n                    или запрошенное лицо не имеет родителя, с переданным parent_id."
                    }
                }
            }
        },
...
        "/api/download/photo": {
            "get": {
                "operationId": "92560dd3a81b0078b01a2c7e6fd2079b",
                "responses": {
                    "200": {
                        "description": "Файл в формате ZIP, содержащий все фотоснимки из родословной."
                    },
                    "204": {
                        "description": "Фотоснимки на сервере отсутствуют."
                    }
                }
            }
        },
...
    }
}